@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 3.0.0
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: '>=3.3.3 <4.0.0'
 
 dependencies:
   flutter:
